Fixes compatibility issues with a few ROMs:
- Alien Soldier: mdZ80 error, non-standard checksum
- Alien Storm: mdZ80 error (same as Alien Soldier)
- World Series Baseball: activated 240-line mode due to an incorrect VDP mode check
- Bubba 'N' Stix: Broken menus and blackscreen due to an incorrect DMA enable check
- Cadash: Non-standard checksum

Auto-fix checksum has been disabled by default, since it seems to cause
more problems than it helps fix nowadays.

gens-qt4: Fixed the initial region code setting and PAL mode.

gens-sdl: Added command line options to set the region code, plus
auto-detection. (Auto-detection order is hard-coded to 0x4812 for now.)

Various cppcheck fixes.

+++ src/libgens/macros/common.h
@@ -100,6 +100,6 @@
  */
 #define ARRAY_SIZE(x) \
 	((int)(((sizeof(x) / sizeof(x[0]))) / \
-	       (size_t)(!(sizeof(x) % sizeof(x[0])))))
+	       (int)(!(sizeof(x) % sizeof(x[0])))))
 
 #endif /* __LIBGENS_MACROS_COMMON_H__ */
